Day 9, Add category engine only, without replacing current categorizer yet

+++ src/main/java/com/moniq/api/categorization/CategorizationResult.java
@@ -1,15 +1,17 @@
 // src/main/java/com/moniq/api/categorization/CategorizationResult.java
 package com.moniq.api.categorization;
 
+import java.math.BigDecimal;
+
 public class CategorizationResult {
     private final String category;
-    private final double confidence;
+    private final BigDecimal  confidence;
 
-    public CategorizationResult(String category, double confidence) {
+    public CategorizationResult(String category, BigDecimal  confidence) {
         this.category = category;
         this.confidence = confidence;
     }
 
     public String getCategory() { return category; }
-    public double getConfidence() { return confidence; }
+    public BigDecimal  getConfidence() { return confidence; }
 }
